feat: add icon slider animation; refact: replace slider code to slider.js
