+++ src/blocks/description/__icon/_type/description__icon_type_gravel.scss
@@ -1,0 +1,3 @@
+.description__icon_type_gravel {
+  background-image: url('../images/description-slider/icon-gravel.svg');
+}
