+++ src/blocks/description/__icon/_type/description__icon_type_highway.scss
@@ -1,0 +1,3 @@
+.description__icon_type_highway {
+  background-image: url('../images/description-slider/icon-higway.svg');
+}
