+++ src/blocks/description/__icon/_type/description__icon_type_tt.scss
@@ -1,0 +1,3 @@
+.description__icon_type_tt {
+  background-image: url('../images/description-slider/icon-tt.svg');
+}
