+++ src/blocks/description/__icon/description__icon.scss
@@ -1,0 +1,24 @@
+.description__icon {
+  width: 50px;
+  height: 100%;
+  background-repeat: no-repeat;
+  background-position: right bottom;
+  transition: skew(25deg, 0);
+  transition: all .5s ease;
+}
+
+@keyframes switch-animation {
+  from {
+    transform: scaleX(1);
+  }
+  50% {
+    transform: scaleX(0);
+  }
+  to {
+    transform: scaleX(1);
+  }
+}
+
+.switch-icon {
+  animation: switch-animation .5s ease 0s;
+}
