+++ src/blocks/slider/slider.scss
@@ -1,4 +1,5 @@
 .slider {
   max-width: 100vw;
   overflow: hidden;
+  position: relative;
 }
