+++ src/index.html
@@ -69,6 +69,9 @@
           <img class="description__image" src="<%=require('./images/description-slider/image-3.png')%>" alt="">
           <img class="description__image" src="<%=require('./images/description-slider/image-4.png')%>" alt="">
         </ul>
+        <div class="description__icon-box">
+          <div class="description__icon description__icon_type_highway"></div>
+        </div>
         <div class="slider__actions">
           <button class="slider__button slider__button_type_prev" type="button" name="prev"></button>
           <button class="slider__button slider__button_type_next" type="button" name="next"></button>
